pyproject.toml
[tool.poetry]
name = "email-invitation-to-contribute"
version = "1.0.0"
description = ""
authors = ["30ème Ciel <admin@30emeciel.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
core = { git = "https://github.com/30emeciel/py-core.git", branch="v2"}
# core = { path = "../py-core", develop = true }
# core = { path = "../py-core", develop = false }
python-box = "^5.3.0"
google-cloud-pubsub = "^2.4.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.17.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
